Refactoring: GeoDataRiverSurvey::ExpandCrosssectionCommand is moved to separated files.
