+++ libs/geodata/riversurvey/riversurvey.pro
@@ -147,6 +147,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_addextensioncommand.h \
            private/geodatariversurvey_changeselectioncommand.h \
            private/geodatariversurvey_deleteriverpathpointcommand.h \
+           private/geodatariversurvey_expandcrosssectioncommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
@@ -195,6 +196,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_addextensioncommand.cpp \
            private/geodatariversurvey_changeselectioncommand.cpp \
            private/geodatariversurvey_deleteriverpathpointcommand.cpp \
+           private/geodatariversurvey_expandcrosssectioncommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.cpp \
            private/geodatariversurveycrosssectionwindow_riversurveytabledelegate.cpp \
